backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
duckduckgo-search
trafilatura
python-dotenv
python-multipart